Cleanup: str_to_datetime moved to utils & add docstring
